Updated comment UI and created share dialog
Made the UI of the comments section better by improving the design as well as the design when the comment belongs to the logged in user. added a refresh button. Create the share dialog template which allows for a custom message. Can be for either reviews or comments and added the share buttons (Copy to clipboard,X,Facebook,Reddit). Next: addfunctionality to the share buttons and also enable sharing for reviews when the share dialog is complete