Added in logic to auto delete comment if it reaches 5 reports
When reporting a comment, if that comments reaches 5 reports it will have its IsDeleted flag set to true. These comments will be able to be reviewed later down the line by an admin to determine if the reports are valid. Stopped all actions from being performed if the comment is deleted