Altered server-side to store userId
UserId now does not need to be sent with every request except 1 (ViewUserSummary on the ViewUserPage). It stores the UserId on the server when the user logs in either via manual login or auto-login. This id is then used instead of sending one from the client. The UserId is cleared when a user logs out or closes their browser (instance closed so clears)